+++ src/mlia/cli/options.py
@@ -62,15 +62,17 @@ def add_multi_optimization_options(parser: argparse.ArgumentParser) -> None:
 
 def add_optional_tflite_model_options(parser: argparse.ArgumentParser) -> None:
     """Add optional model specific options."""
-    model_group = parser.add_argument_group("TFLite model options")
+    model_group = parser.add_argument_group("TensorFlow Lite model options")
     # make model parameter optional
-    model_group.add_argument("model", nargs="?", help="TFLite model (optional)")
+    model_group.add_argument(
+        "model", nargs="?", help="TensorFlow Lite model (optional)"
+    )
 
 
 def add_tflite_model_options(parser: argparse.ArgumentParser) -> None:
     """Add model specific options."""
-    model_group = parser.add_argument_group("TFLite model options")
-    model_group.add_argument("model", help="TFLite model")
+    model_group = parser.add_argument_group("TensorFlow Lite model options")
+    model_group.add_argument("model", help="TensorFlow Lite model")
 
 
 def add_output_options(parser: argparse.ArgumentParser) -> None: