+++ tests/test_nn_tensorflow_config.py
@@ -14,7 +14,7 @@ from mlia.nn.tensorflow.config import TfModel
 
 
 def test_convert_keras_to_tflite(tmp_path: Path, test_keras_model: Path) -> None:
-    """Test Keras to TFLite conversion."""
+    """Test Keras to TensorFlow Lite conversion."""
     keras_model = KerasModel(test_keras_model)
 
     tflite_model_path = tmp_path / "test.tflite"
@@ -25,7 +25,7 @@ def test_convert_keras_to_tflite(tmp_path: Path, test_keras_model: Path) -> None
 
 
 def test_convert_tf_to_tflite(tmp_path: Path, test_tf_model: Path) -> None:
-    """Test TensorFlow saved model to TFLite conversion."""
+    """Test TensorFlow saved model to TensorFlow Lite conversion."""
     tf_model = TfModel(test_tf_model)
 
     tflite_model_path = tmp_path / "test.tflite"
@@ -47,7 +47,8 @@ def test_convert_tf_to_tflite(tmp_path: Path, test_tf_model: Path) -> None:
             pytest.raises(
                 Exception,
                 match="The input model format is not supported"
-                r"\(supported formats: TFLite, Keras, TensorFlow saved model\)!",
+                r"\(supported formats: TensorFlow Lite, Keras, "
+                r"TensorFlow saved model\)!",
             ),
         ),
     ],
@@ -55,7 +56,7 @@ def test_convert_tf_to_tflite(tmp_path: Path, test_tf_model: Path) -> None:
 def test_get_model_file(
     model_path: str, expected_type: type, expected_error: Any
 ) -> None:
-    """Test TFLite model type."""
+    """Test TensorFlow Lite model type."""
     with expected_error:
         model = get_model(model_path)
         assert isinstance(model, expected_type)
@@ -67,6 +68,6 @@ def test_get_model_file(
 def test_get_model_dir(
     test_models_path: Path, model_path: str, expected_type: type
 ) -> None:
-    """Test TFLite model type."""
+    """Test TensorFlow Lite model type."""
     model = get_model(str(test_models_path / model_path))
     assert isinstance(model, expected_type)